requirements.txt
ultralytics
opencv-python-headless
numpy
Pillow
streamlit
